Controller finalizers, conditions and labels fixes
AO-48 : Add finializers to trials and experiments to facilitate a more orderly deletion; including communication with the server

Conditions: instead of a simple “failed” boolean state, support conditions similar to how Kubernetes handles jobs and pods

Labels: Fix a significant bug where the default behavior created jobs using a labeling scheme that caused subsequent trials to adopt previous jobs and immediately mis-report status based on the previous job. The new scheme is more consistent and allows selectors to default to the labels provided in the corresponding templates; the fall back also uses different labels for trials and experiments to avoid the aforementioned bug.