Separate tests from Docker build
Kubebuilder 2.0 made the Docker build depend on the tests, but that just isn't a reasonable expectation in this project. For one, it makes the Circle build logs impossible to follow (since Go modules are downloaded twice); it also makes testing the setup tools Docker image almost impossible due to unnecessary Go testing.