Add a TTL to trials
This changes the existing behavior: instead automatically cleaning up trials, the behavior needs to be opted into. This makes more sense with a normal workflow as during the initial testing phases experiment authors are more likely to require the trials be preserved for debugging.

This does not effect the "owner" relationship: deleting and experiment still cascades to the owned trials. This behavior is not enforced if the experiment controller is not in use. The legacy behavior can be achieved by explicitly setting the TTL to 0.