Accept a "forbidden" error as a stand-in for a setup delete job
If creating a setup delete job fails with a "forbidden" error, assume it is because the namespace is being torn down and new objects are not being accepted. Since deleting the namespace should remove any of the objects the setup delete job would have deleted, just update the trial to indicate setup deletion has completed.