Attempt to prevent concurrent trial additions to the same namespace
This change assumes that multiple attempts are made to add a trial to the same namespace using the same copy of the cache (i.e. both trials and namespaces): by changing the namespace to include a random value the idea is that the API server will report a conflict and by the time the re-queue is processed the cache will be (more) consistent.