refactor: Allow ReadExperiments to flex on input filetype
This changes ReadExperiments to use yaml.NewYamlOrJsonDecoder so that we can
remove the filename suffix codepath. This allows us to use a tempfile for testing
( which wouldnt contain a suffix ) as well as removes some of the fragility of
file type/media type detection.

As part of this change, we also change the function signature to a singleton. We
dont support multiple experiments defined in a single file, so we shouldnt return
an experiment list.

Signed-off-by: Brad Beam <brad.beam@carbonrelay.com>