refactor: Export name struct and rename to Identifer
We are going to be reusing the name struct for the export lookup,
so we need to be able to explicitly reference this type to pass
it around as arguments.
The rename was done to reduce the duplicaion of Name.Name and help
put a bit more meaning/context with the struct.

Signed-off-by: Brad Beam <brad.beam@carbonrelay.com>