Fix secret generation default client name
The configuration has not been loaded yet when client name needs to be initialized (and in the case of --help it never will be).

Directly exec'ing the command removes the dependency on loading the configuration.

Also, added a KUBECONFIG when generating documentation to prevent the local environment from leaking into the published documentation.