refactor: Consolidate prometheus deployment into a single pod
This consolidates kube state metrics, prometheus server, and push gateway
into a single prometheus pod/deployment. There's not a strong use case for
keeping them separate and this should simplify the scraping of these endpoints
in the prometheus configuration since we'll only scrape localhost instead of
needing to account for the generated service names.

Signed-off-by: Brad Beam <brad.beam@carbonrelay.com>