fix: Remove CrashBackoffLoop from pod checks
This fixes an issue where we might misidentify a failing pod and fail a trial.

One scenario happens when a previous pod is in a CrashLoopBackOff state. If a new
pod does not come up healthy in time we check the pod status, we would trigger
this as a hard failure and cause the trial to fail.

Another scenario happens when a pod relies on the platform to handle dependencies
being available. If PodA checks for ServiceBs availability during startup and
terminates if ServiceB is unavailable, kubernetes will mark this pod as
being stuck in a CrashLoopBackOff.

Signed-off-by: Brad Beam <brad.beam@stormforge.io>