refactor: move repeated code into a function
This introduces a generateAndWait function for running our integration tests
to help DRY the tests here.

This also adds in some additional waits to ensure that we are in fact ready to
start up the secondary experiment. I think we were hitting an issue with
deleting resources associated with the first experiment ( deleting the service
account before the trial deletion job finished ).

Signed-off-by: Brad Beam <brad.beam@stormforge.io>