fix: Ensure prometheus resources are removed before creation
This fixes ( works around? ) an issue where the prometheus resources
dont get properly cleaned up. This typically causes the service account
to be deleted prior to the deletion task which causes a silent failure.

This is most commonly reproducable when the output of generate experiment
is piped into `kubectl apply -f -` and `kubectl delete -f -`.

Signed-off-by: Brad Beam <brad.beam@stormforge.io>