fix: Readd support for sanitizing application labels
We have a bit of a mix and match for label support. We currently wrap the prom labels
with kube label selectors. This means that the prom label needs to satisfy a DNS1123
format ( to make kube happy ) and also satisfy a prometheus label format ( a-zA-Z0-9_ ).
So this leaves us with only supporting a `a-z0-9` string for the label key.

Valid/Recommended kubernetes labels such as `app.kubernetes.io/name` are accepted but do
not actually work because prometheus doesnt care for the `.` and `/` characters resulting
in a cryptic error message like:

```
bad_data: invalid parameter "query": 7:64: parse error: unexpected character inside braces: '.'
```

Signed-off-by: Brad Beam <brad.beam@stormforge.io>