Fix `set config` for controllers with dot names
Previously if a controller had a dot character "." in the name, it
wasn't possible to set configs on that controller.

This commit fixes the issue by using optimize-go 0.0.28, which is
assumed to support escaping dot characters in path element names with a
backslash. E.g. "one\.two" is a single path element.